refact : feature/detail 구조 분리 및 계획
- detail/ui 내부로 content를 분리
- ResizeBook ? 기능을 위젯으로 분리 계획중
- CoverLetterDetailLayout의 경우
- page레이어에서 detail, preivew 조합으로 관리할 계획중
