+++ client/src/features/coverletter/detail/index.ts
@@ -1,0 +1,1 @@
+export { default as CoverLetterDetailLayout } from './ui/CoverLetterDetailLayout';
