+++ client/src/features/coverletter/detail/ui/content/CoverLetterDetail.tsx
@@ -1,0 +1,55 @@
+import { styled } from 'styled-components';
+
+import { coverLetterQueries } from '@/src/entities/coverLetter';
+
+import { styleMixin, V } from '@/src/shared/styles';
+
+import CoverLetterContent from './CoverLetterContent';
+import CoverLetterHeader from './CoverLetterHeader';
+
+type Props = {
+  essayId: number;
+};
+
+export default function CoverLetterDetail({ essayId }: Props) {
+  const { data: coverLetterData } = coverLetterQueries.useFetchDetailCoverLetter(essayId);
+  const { qnaList } = coverLetterData;
+  return (
+    <CoverLetterContainer>
+      <HeaderWrapper>
+        <CoverLetterHeader
+          title={coverLetterData.title}
+          questionLen={qnaList.length}
+        />
+      </HeaderWrapper>
+      <ContentWrapper>
+        {qnaList.map((item, index) => (
+          <CoverLetterContent key={item.qnaId} index={index} {...item} />
+        ))}
+      </ContentWrapper>
+    </CoverLetterContainer>
+  );
+}
+
+const CoverLetterContainer = styled.div`
+  width: 100%;
+  height: 100%;
+  border: 1px solid ${(props) => props.theme.colors.mainLine};
+  background-color: ${(props) => props.theme.colors.primary};
+  box-shadow: ${V.boxShadow2};
+  overflow-y: auto;
+  ${styleMixin.ScrollBar};
+  padding-left: 2rem;
+`;
+
+const HeaderWrapper = styled.div`
+  padding: 2rem 2rem 1.5rem 0;
+  border-bottom: 1px solid ${(props) => props.theme.colors.mainLine};
+  margin-bottom: 2rem;
+`;
+
+const ContentWrapper = styled.div`
+  ${styleMixin.Column('flex-start', 'flex-start')}
+  padding: 0 2rem 3rem 0;
+  gap: 3rem;
+`;
