+++ client/src/features/coverletter/detail/ui/content/index.ts
@@ -1,0 +1,1 @@
+export { default as CoverLetterDetail } from './CoverLetterDetail';
